Remove unicode characters in the document example
Lua strings are latin-1. So '“' and '”' needs to be replaced by '"' in
the document example.

TODO: evaluate whether UTF-8 strings should be used in Lua. For this
particular case, we can use online formatting as double-quotes are made
with the \quotation in ConTeXt. OTOH, markdown understands texts
sourrounded by '"' as double-quoted text.