Refactor file loading and change file formats
* Loading text content from a document file had multiple issues that are
now correct. Here are the details:

** Change the delimiter after the end key from ':' to ':=', because a
new line of text that contains a ':' will be wrongly interpreted as
'key: value'. This also lead to simplifications in the code as it
doesn't need anymore to find out which key is an end key. Note that I
didn't choose a plain '=' as delimiter, because, once implemented, the
declare key will contain lua variable assignments which can be confused
by the reader as a end key assignment. By the way ':=' is how to assign
variables in some programming languages such as Ada and originates from
mathematics.

** Detect the type of a new line (value or key) based on its
indentation, instead on whether it looks like a key or not.

** In document file, escape texts with "", to avoid unwanted trimming and
to simply detection of blank values, and fix wrong indentation of texts
when they are multiline.

** Fix issue with comment detection that removes text in a text block if
it begins by '--'

* Fix visibility issues with classes derivating from
Abstract_Loadable_Object, as some have the assign and process_key public
instead of protected.

* Set version to 0.0.0 for all files, as the file formats are not
stabilised yet.