Refactor definition files, document classes and styles
Document classes are now the combination of a generic class and a backend class. This avoids numerous duplications, when a new class only consists on a different layout.

Further changes in files structure.

Removal of the properties key and enable intermediate keys to be placed right before end keys. The previous limitation was error-prone and can add unecessarily complexity in the definition files.

Refactor exceptions