Fix focus management
Default behaviour after a widget insertion is to insert the widget at
the end of the focus chain with no respect where it was inserted in the
document area. I was struggling with fixing that behaviour as fixing the
focus chain at one point may mess up another part of it. So I created a
focus manager, to which will only the relevant widget be registered, and
I overrided focusNextPrevChild to ask the focus manager to focus on a
sibling widget.

Some smaller changes:

* Fix wrong level for children widgets

* After a text edit gets the focus, ensure that it will be visible on the
scroll area