Change behaviour on insertion
Insertion is now made to the next available position on the current
children widget, if available, otherwise on the first position of the
next children widget. Moreover the style of the new block corresponds
to the default style instead of the style of the sibling.

This is in line with the behavior of MS Word and would also be less
confusing for the user.